+++ src/dataStructures/tree/binarySearchTree/leetcode/__tests__/isValid.test.ts
@@ -13,6 +13,17 @@ test('isValid', () => {
     .toBeTruthy();
 
   expect(isValid(serializeLevelOrder([
+    2,
+    1,
+    3,
+    null,
+    null,
+    4,
+    null,
+  ])))
+    .toBeFalsy();
+
+  expect(isValid(serializeLevelOrder([
     5,
     1,
     4,